+++ src/screens/WalletScreen/components/Header/Header.types.ts
@@ -3,4 +3,5 @@ import { GestureResponderEvent } from 'react-native';
 export interface HeaderProps {
 	onSettingsPress: (event: GestureResponderEvent) => void;
 	onCopyPress: (event: GestureResponderEvent) => void;
+	accountName: string;
 }
