+++ src/screens/WelcomeScreen/ImportWalletModal/useImportWallet.hooks.ts
@@ -3,12 +3,9 @@ import { captureException } from '@sentry/react-native';
 import { useState } from '@hookstate/core';
 import { useFormProgress, useNavigation } from '@hooks';
 import { Keyboard } from 'react-native';
-import crypto from 'crypto';
 import { globalWalletState, walletState } from '@src/stores/WalletStore';
 import { restoreWalletByMnemonic } from '@src/model/wallet';
-import Intercom from '@intercom/intercom-react-native';
 import Logger from '@utils/logger';
-import { INTERCOM_KEY } from '@env';
 
 interface UseImportWalletModalProps {
 	onImportFinished: () => void;
@@ -33,15 +30,6 @@ export const useImportWalletModal = ({ onImportFinished, visible, onDismiss }: U
 				state.set(await walletState(wallet));
 				setImporting(false);
 				onImportFinished();
-
-				// Register user on Intercom
-				const intercomKey = INTERCOM_KEY || process.env.INTERCOM_KEY;
-				const hmac = crypto.createHmac('sha256', intercomKey!);
-				hmac.update(wallet.address);
-				const sign = hmac.digest('hex');
-				Intercom.setUserHash(sign);
-				Intercom.registerIdentifiedUser({ userId: wallet.address });
-				//
 			} catch (error) {
 				setImporting(false);
 				Logger.error('Invalid seed phrase or primary key');
